w3c-reports/12092014/mac-wk2/runned-tests/html/infrastructure/urls/resolving-urls/query-encoding/utf-16be-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:49001/echo-query?%C3%A5' failed: WebSocket is closed before the connection is established.
CONSOLE MESSAGE: line 1: SyntaxError: Invalid character '\u8224'
CONSOLE MESSAGE: line 1: SyntaxError: Invalid character '\u8224'
CONSOLE MESSAGE: line 1: SyntaxError: Invalid character '\u8224'
FAIL: Timed out waiting for notifyDone to be called
